requirements.txt
torch
torchvision
pytorch-lightning
torchmetrics
pandas
scikit-learn
pillow
